# Improve performance of the tournament page